chore: move NamedLocation, GpsLocation, useNominatim and LocationCoordinateSelect to core package
