+++ packages/core/src/composables/index.ts
@@ -5,13 +5,15 @@ import {
 import { useDate } from '@core/composables/useDate'
 import { useFetchState } from '@core/composables/useFetchState'
 import { useInterval } from '@core/composables/useInterval'
+import { useNominatim } from '@core/composables/useNominatim'
 import { useNumber } from '@core/composables/useNumber'
 
 export {
-  useBrowserPermissions,
   PermissionState,
+  useBrowserPermissions,
   useDate,
   useFetchState,
   useInterval,
+  useNominatim,
   useNumber
 }
