+++ packages/core/src/inputs/LocationCoordinateSelect/LocationCoordinateSelect.stories.ts
