+++ packages/core/src/inputs/index.ts
@@ -5,11 +5,12 @@ import BaseSelect from '@core/inputs/BaseSelect/BaseSelect.vue'
 import CheckboxInput from '@core/inputs/CheckboxInput/CheckboxInput.vue'
 import DateTimeInput from '@core/inputs/DateTimeInput/DateTimeInput.vue'
 import ImageInputWrapper from '@core/inputs/ImpageInputWrapper/ImageInputWrapper.vue'
+import LocationCoordinateSelect from '@core/inputs/LocationCoordinateSelect/LocationCoordinateSelect.vue'
 import NumberInput from '@core/inputs/NumberInput/NumberInput.vue'
 import PreviewImageInput from '@core/inputs/PreviewImageInput/PreviewImageInput.vue'
 import SelectInput from '@core/inputs/SelectInput/SelectInput.vue'
-import UploadImageFAB from '@core/inputs/UploadImageFAB/UploadImageFAB.vue'
 import SliderToggleInput from '@core/inputs/SliderToggleInput/SliderToggleInput.vue'
+import UploadImageFAB from '@core/inputs/UploadImageFAB/UploadImageFAB.vue'
 
 export {
   BaseInput,
@@ -19,9 +20,10 @@ export {
   CheckboxInput,
   DateTimeInput,
   ImageInputWrapper,
+  LocationCoordinateSelect,
   NumberInput,
   PreviewImageInput,
   SelectInput,
-  UploadImageFAB,
-  SliderToggleInput
+  SliderToggleInput,
+  UploadImageFAB
 }
