+++ packages/core/src/models/gpsLocation.ts
