+++ packages/core/src/models/index.ts
@@ -1,3 +1,7 @@
 import { SelectOption } from '@core/inputs/BaseSelect/selectOption'
+import { GpsLocation } from '@core/models/gpsLocation'
+import { NamedLocation } from '@core/models/namedLocation'
+import { type NominatimResult } from '@core/models/nominatimResult'
 
-export { SelectOption }
+export { SelectOption, GpsLocation, NamedLocation }
+export type { NominatimResult }
