+++ packages/core/src/models/nominatimResult.ts
