+++ packages/gridstack/src/components/widgets/weather/widgets/WeatherForcastWidget/config.ts
@@ -4,8 +4,7 @@ import {
 } from '@/components/widgets/weather/widgets/WeatherForcastWidget/settings'
 import { WidgetComponentWrapper } from '@/models/widgetComponentWrapper'
 import WeatherForecastWidget from '@/components/widgets/weather/widgets/WeatherForcastWidget/WeatherForecastWidget.vue'
-import { NamedLocation } from '@/models/namedLocation'
-import { GpsLocation } from '@/models/gpsLocation'
+import { NamedLocation, GpsLocation } from '@lion5/component-library'
 
 export const WEATHER_FORECAST_WIDGET_WRAPPER = new WidgetComponentWrapper(
   'Wettervorhersage',
