+++ packages/gridstack/src/components/widgets/weather/widgets/WeatherWidget/WeatherWidget.vue
@@ -44,7 +44,7 @@ import WindDirection from '@/components/widgets/weather/components/WindDirection
 import BrightSkySources from '@/components/widgets/weather/components/BrightSkySources.vue'
 import { WidgetSettings } from '@/models/widgetConfiguration'
 import { useWidget } from '@/composables/useWidget'
-import { NamedLocation } from '@/models/namedLocation'
+import { NamedLocation } from '@lion5/component-library'
 
 const props = defineProps<{
   settings: WidgetSettings
