+++ packages/leaflet/src/components/inputs/index.ts
@@ -1,4 +1,3 @@
 import GeolocationInput from '@leaflet/components/inputs/GeolocationInput/GeolocationInput.vue'
-import LocationCoordinateSelect from '@leaflet/components/inputs/LocationCoordinateSelect/LocationCoordinateSelect.vue'
 
-export { GeolocationInput, LocationCoordinateSelect }
+export { GeolocationInput }
