+++ packages/leaflet/src/composables/index.ts
@@ -2,7 +2,6 @@ import { useGeoJSON } from '@leaflet/composables/useGeoJSON'
 import { useLeafletButtonControl } from '@leaflet/composables/useLeafletButtonControl'
 import { useLeafletMap } from '@leaflet/composables/useLeafletMap'
 import { useLeafletMarker } from '@leaflet/composables/useLeafletMarker'
-import { useNominatim } from '@leaflet/composables/useNominatim'
 import { useShape } from '@leaflet/composables/useShape'
 
 export {
@@ -10,6 +9,5 @@ export {
   useLeafletButtonControl,
   useLeafletMap,
   useLeafletMarker,
-  useNominatim,
   useShape
 }
