+++ packages/leaflet/src/models/index.ts
@@ -4,7 +4,6 @@ import { ButtonControl } from '@leaflet/models/leafletButtonControl'
 import type { ApiGpsLocation } from '@leaflet/models/gpsLocation'
 import { GpsLocation } from '@leaflet/models/gpsLocation'
 import { NamedLocation } from '@leaflet/models/namedLocation'
-import { NominatimResult } from '@leaflet/models/nominatimResult'
 
 export {
   LocalResource,
@@ -15,4 +14,3 @@ export {
   NamedLocation,
   ApiGpsLocation
 }
-export type { NominatimResult }
